third_party/blink/web_tests/platform/generic/media/media-fragments/TC0037-expected.txt


Title: TC0037
Fragment: 't=3%2C7'
Comment: The media is requested from 3 to 7 seconds. Note: %2C is equivalent to ,.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '3') OK
RUN(video.play())
EVENT(pause)
EXPECTED (video.currentTime - fragmentEndTime <= '0.5') OK
END OF TEST
